patch-1.04A bug fix and balance
Didn't use the equivalent of an else statement whilst using ternary operator hence it would throw an error when generating a pick greater than 10.  I have made it so that the highest level monsters don't appear as much but at the same time reduce chance of collecting a herb after a battle.